traces: reorganize functions
