+++ web/src/features/traces/components/TraceSearchList.tsx
@@ -11,7 +11,7 @@ import { useTraceData } from "@/src/features/traces/contexts/TraceDataContext";
 import { useSearch } from "@/src/features/traces/contexts/SearchContext";
 import { useSelection } from "@/src/features/traces/contexts/SelectionContext";
 import { useHandlePrefetchObservation } from "@/src/features/traces/hooks/useHandlePrefetchObservation";
-import { useDesktopLayoutContextOptional } from "./_layout/TraceLayoutDesktop";
+import { useDesktopLayoutContextOptional } from "./TraceLayoutDesktop";
 import { usePostHogClientCapture } from "@/src/features/posthog-analytics/usePostHogClientCapture";
 import { useTraceAnalyticsDimensions } from "@/src/features/traces/hooks/useTraceAnalyticsDimensions";
 import { VirtualizedList } from "./_shared/VirtualizedList";
