+++ web/src/features/traces/components/TraceTree.tsx
@@ -18,8 +18,8 @@ import { useTraceData } from "@/src/features/traces/contexts/TraceDataContext";
 import { useSelection } from "@/src/features/traces/contexts/SelectionContext";
 import { useIsObservationActive } from "@/src/features/traces/contexts/PlayheadContext";
 import { useHandlePrefetchObservation } from "@/src/features/traces/hooks/useHandlePrefetchObservation";
-import { useDesktopLayoutContextOptional } from "./_layout/TraceLayoutDesktop";
-import { useMobileLayoutContextOptional } from "./_layout/TraceLayoutMobile";
+import { useDesktopLayoutContextOptional } from "./TraceLayoutDesktop";
+import { useMobileLayoutContextOptional } from "./TraceLayoutMobile";
 import { type TreeNode } from "../fns/types";
 import { cn } from "@/src/utils/tailwind";
 import type Decimal from "decimal.js";
